[zsh] Put asdf configuration again in zshrc
Turns out macOS has these two things called `/etc/paths` and `/etc/paths.d/*` that literally prepend themselves _after_ .zshenv, kinda defeating the whole fucking purpose of having one for that end (JOY 🎉). I was fucking with it before because my `.zprofile` was overwriting the $PATH...

This is fucking great.

Looks like asdf's script is just making sure the PATH is right (actually removing it from the PATH and prepending it if not!), but I still need it in .zshenv because that's what vim uses behind the scenes for its commands (fun fact, the PATH doesn't get messed up in that scenario).

Fucking end me now.